requirements.txt
fastapi
uvicorn
python-socketio